+++ src/Sulu/Bundle/WebsiteBundle/Cache/CacheClearer.php
@@ -12,6 +12,8 @@
 namespace Sulu\Bundle\WebsiteBundle\Cache;
 
 use Sulu\Bundle\HttpCacheBundle\Cache\CacheManager;
+use Sulu\Bundle\WebsiteBundle\Events;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\HttpFoundation\RequestStack;
 
@@ -50,11 +52,17 @@ class CacheClearer implements CacheClearerInterface
      */
     private $cacheManager;
 
+    /**
+     * @var EventDispatcherInterface
+     */
+    private $eventDispatcher;
+
     public function __construct(
         Filesystem $filesystem,
         $kernelEnvironment,
         $kernelRootDir,
         RequestStack $requestStack,
+        EventDispatcherInterface $eventDispatcher,
         $varDir = null,
         ?CacheManager $cacheManager
     ) {
@@ -64,6 +72,7 @@ class CacheClearer implements CacheClearerInterface
         $this->varDir = $varDir;
         $this->cacheManager = $cacheManager;
         $this->requestStack = $requestStack;
+        $this->eventDispatcher = $eventDispatcher;
     }
 
     /**
@@ -79,6 +88,8 @@ class CacheClearer implements CacheClearerInterface
 
             $this->cacheManager->invalidateDomain($request->getHost());
 
+            $this->eventDispatcher->dispatch(Events::CACHE_CLEAR);
+
             return;
         }
 
@@ -91,5 +102,7 @@ class CacheClearer implements CacheClearerInterface
         if ($this->filesystem->exists($path)) {
             $this->filesystem->remove($path);
         }
+
+        $this->eventDispatcher->dispatch(Events::CACHE_CLEAR);
     }
 }