+++ src/Sulu/Component/PHPCR/PathCleanup.php
@@ -74,10 +74,13 @@ class PathCleanup implements PathCleanupInterface
 
         // Inspired by ZOOLU
         // delete problematic characters
-        $clean = str_replace('%2F', '/', urlencode(preg_replace('/([^A-za-z0-9 -_\/])/', '', $clean)));
+        $clean = str_replace('%2F', '/', urlencode(preg_replace('/([^A-za-z0-9\s\-_\/])/', '', $clean)));
 
         // replace multiple dash with one
         $clean = preg_replace('/([-]+)/', '-', $clean);
+
+        // remove dash before slash
+        $clean = preg_replace('/[-]+\//', '/', $clean);
 
         // remove dash after slash
         $clean = preg_replace('/\/[-]+/', '/', $clean);